requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-admin-interface==0.28.1
django-colorfield==0.11.0
django-ninja==1.1.0
idna==3.6
pillow==10.2.0
psycopg-binary==3.1.16
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-slugify==8.0.1
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
